feat:pratica-05 -listar a placa de todos os veiculos

+++ src/storage03/pratica05/src/main/java/com/emily/pratica05/model/Veiculo.java
@@ -16,7 +16,7 @@ public class Veiculo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id_vehicle;
+    private int id_vehicle;
 
     private String board;
